+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day35/Solution188_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan01.day35;
+package com.noob.algorithm.daily.plan01.archive.day35;
 
 /**
  * 🔴 188 买卖股票的最佳时机IV - https://leetcode.cn/problems/best-time-to-buy-and-sell-stock-iv/description/
