+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day35/Solution714_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan01.day35;
+package com.noob.algorithm.daily.plan01.archive.day35;
 
 /**
  * 🟡 714 买卖股票的最佳时机含手续费 - https://leetcode.cn/problems/best-time-to-buy-and-sell-stock-with-transaction-fee/description/
