+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day35/Solution912_02.java
@@ -1,9 +1,9 @@
-package com.noob.algorithm.daily.plan01.day35;
+package com.noob.algorithm.daily.plan01.archive.day35;
 
 /**
  * 🟡 912 手撕快速排序（排序数组） - https://leetcode.cn/problems/sort-an-array/description/
  */
-public class Solution912_01 {
+public class Solution912_02 {
 
     /**
      * 快速排序思路：
@@ -36,21 +36,21 @@ public class Solution912_01 {
     private int partition(int[] nums, int left, int right) {
         // ① 选择第1个元素作为base
         int base = nums[left];
-        int cur = left + 1; // 定义当前分界位置
+        int cur = left; // 定义当前分界位置
 
         // ② 遍历元素，将base放在合适的位置
         for (int i = left + 1; i <= right; i++) {
             // 如果遇到比base小的元素则将其交换到前面来
             if (base > nums[i]) {
+                cur++; // 分界位置往前移动一位（相当于指向交换位置，每次都是要执行交换前才空出位置）
                 swap(nums, i, cur); // 执行交换
-                cur++; // 分界位置继续往前移动一位
             }
         }
 
-        // 最后将base交换到cur-1指向位置
-        swap(nums, left, cur - 1); // 原base位置为left，指向交换位置cur-1
+        // 最后将base交换到cur指向位置
+        swap(nums, left, cur); // 原base位置为left，指向交换位置cur
         // 返回交换后的base的位置
-        return cur - 1;
+        return cur;
     }
 
     // 交换元素
