+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day36/Solution300_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan01.day36;
+package com.noob.algorithm.daily.plan01.archive.day36;
 
 import java.util.Arrays;
