+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day37/Solution053_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan01.day37;
+package com.noob.algorithm.daily.plan01.archive.day37;
 
 import java.util.Arrays;
